requirements.txt
PyAudio==0.2.14
openai-whisper==20231117
openai==0.28.1
python-dotenv==1.0.1
whisper
google-cloud-texttospeech
SpeechRecognition
flask
flask-socketio
python-socketio 